refactor: 태그, 역량 키워드 애그리거트 분리(#46)
