+++ Domain-Module/src/main/kotlin/com/bamyanggang/domainmodule/domain/strongpoint/aggregate/StrongPoint.kt
@@ -1,6 +1,6 @@
-package com.bamyanggang.domainmodule.domain.experience.aggregate
+package com.bamyanggang.domainmodule.domain.strongpoint.aggregate
 
-import com.bamyanggang.domainmodule.common.entity.AggregateRoot
+import com.bamyanggang.domainmodule.common.entity.DomainEntity
 import com.example.uuid.UuidCreator
 import java.util.*
 
@@ -8,7 +8,7 @@ data class StrongPoint(
     override val id: UUID = UuidCreator.create(),
     val name : String,
     val userId : UUID?,
-) : AggregateRoot{
+) : DomainEntity{
     companion object {
         fun create(name: String, userId: UUID?): StrongPoint {
             return StrongPoint(UuidCreator.create(), name, userId)
